Add files via upload
I added the DB using sqllite. Getting the top 5 scores was not working with a db query, so I stored the DB in an Arraylist and let the Arraylist sort the data based on highest to lowest scores. Not efficient I know but the SQLite query was not sorting data properly